assets/vector_store_query/requirements.txt
chromadb>=0.4.0
pinecone-client>=3.0.0
qdrant-client>=1.7.0
faiss-cpu>=1.7.4
pandas>=2.0.0
